ASoC: rockchip: i2s: mark FIFOLR register volatile
The FIFO levels will be updated by the device so they should not be
cached.  This register is not accessed by the driver at the moment, so
this has no functional change except for the regmap access in debugfs
but let's fix it while we remember so as to avoid surprises later.

Signed-off-by: John Keeping <john@metanate.com>