ASoC: rt5677: Add ACPI support
The Chromebook Pixel 2015 uses this codec with the ACPI ID RT5677CE, but
does not use the standard DT property names so add a new function to
parse the codec properties from these ACPI properties.

Also, the GPIOs are only available by index, so we need to register a
mapping to allow machine drivers to access the GPIOs by name.

Signed-off-by: John Keeping <john@metanate.com>

---
Changes in v2:
- Add explicit values to ACPI GPIO enum
- Add new function for reading ACPI properties, leave the existing
  rt5677_read_device_properties() as it was
- Add entry to i2c_device_id table instead of adding acpi_driver_id