spi: spi-ep93xx: fix cs_change set on last transfer
The cs_change flag has two meanings depending on whether the transfer is
the last one in a message or not.  When the flag is set on the final
transfer in a message, the chip select must be left active until the
next message needs to be sent to the device.

Signed-off-by: John Keeping <john@metanate.com>